Pridane prazdne metody na pracu v triede obchod

+++ Statistiky.h
@@ -4,16 +4,16 @@
 
 #ifndef POS_UHLAR_DVORSTIAK_SEMESTRALNA_PRACA_BOJOVNICI_STATISTIKY_H
 #define POS_UHLAR_DVORSTIAK_SEMESTRALNA_PRACA_BOJOVNICI_STATISTIKY_H
-
+#include <iostream>
 
 class Statistiky {
 
 public:
-    int dajZivotyHodnota(int uroven, int hranica);
-    int dajBrnenieHodnota(int uroven, int hranica);
-    int dajPoskodenieHodnota(int uroven, int hranica);
-    int dajUnikHodnota(int uroven, int hranica);
-    int dajRychlostUtokuHodnota(int uroven, int hranica);
+    double dajZivotyHodnota(int uroven, int hranica);
+    double dajBrnenieHodnota(int uroven, int hranica);
+    double dajPoskodenieHodnota(int uroven, int hranica);
+    double dajUnikHodnota(int uroven, int hranica);
+    double dajRychlostUtokuHodnota(int uroven, int hranica);
     int dajPocetBojovnikov(int uroven, int hranica);
     void vypisStatistiky(int statistika , int uroven);
